move std-map.hpp
(std-map) Fix include guard

* apply pre-commit formatting
* rename namespace
